Handle a GnuCash limitation related to Stock purchase registers
The "Edit exchange rate" dialog does not show up on stock purchase registers. When I try to open it
using Right click > Edit exchange rate, I see the error message `This register does not support
editing exchange rates." So, I guess the exchange rate can only be changed from the other side of
the stock purchase transaction. This is easy to forget though. So, we should fall back to the
transaction-level value (which will always balance) when quantity is 0 but value is non-0 (which
indicates that the exchange rate was incorrect)